server/requirements.txt
Jinja2==3.1.3
loguru==0.7.2
pyserial==3.5
uvicorn==0.29.0
fastapi==0.110.0
deepface==0.0.89
tf-keras==2.16.0
python-dotenv==1.0.1
opencv-python==4.9.0.80